requirements.txt
fastapi
uvicorn
requests
pydantic
jinja2
python-dateutil
jobsparser
